% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_data}
\alias{format_data}
\title{Format data for \pkg{TMB} calculations.}
\usage{
format_data(y, X, id)
}
\arguments{
\item{y}{Response vector for all subjects.  A vector of length \code{n}.}

\item{X}{Covariate matrix from all subjects.  A matrix of size \code{n x p}.}

\item{id}{Subject identifiers.  A vector of length \code{n} consisting of \code{nsub <= n} distinct elements.  If missing default to \code{rep(1, n)}, i.e., only one subject.}
}
\value{
A list with the following elements:
\describe{
  \item{\code{y}}{The response vector \code{y}, reordered such that all observations for a given subject are consecutive, and converted to an \code{n x 1} column matrix.}
  \item{\code{Xtr}}{The transpose of the covariate matrix, reordered the same way as \code{y}, having size \code{p x n}.}
  \item{\code{iStart}}{An integer vector of length \code{nsub}, specifying the starting index of the observations for each subject, using C++ indexing (i.e., starting at zero).}
  \item{\code{nObs}}{An integer vector of length \code{nsub}, specifying the number of observations per subject.}
}
}
\description{
Format data for \pkg{TMB} calculations.
}
