% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lchol_prior.R
\name{lchol_prior}
\alias{lchol_prior}
\title{Scale-invariant variance matrix prior on the log-Cholesky scale.}
\usage{
lchol_prior(Omega)
}
\arguments{
\item{Omega}{Precision matrix parameter.  A matrix of size \code{p x p}, or an array of size \code{p x p x n}.}
}
\value{
A log-prior vector of length \code{n}.
}
\description{
Scale-invariant variance matrix prior on the log-Cholesky scale.
}
\details{
For a given variance (or precision) matrix \code{Omega}, the scale-invariant prior is
\preformatted{
pi(Omega) \propto 1/|Omega|
}
in the sense that the change of variables \code{Psi = A Omega t(A)} results in the prior \code{pi(Psi) = 1/|Psi|}.

On the log-Cholesky scale defined by \code{\link{log_chol}}, the scale-invariant prior is given by
\preformatted{
pi(logC) \propto prod( diag(chol(Omega))^(p-(1:p)) )
}
}
