% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_chol.R
\name{log_chol}
\alias{log_chol}
\alias{ilog_chol}
\title{Compute the log-Cholesky decomposition and its inverse.}
\usage{
log_chol(V)

ilog_chol(logC)
}
\arguments{
\item{V}{A variance matrix of size \code{p x p}.}

\item{logC}{The log-Cholesky decomposition, represented as a vector of length \code{p*(p+1)/2} (see \strong{Details}).}
}
\value{
For \code{log_chol}, the log-Cholesky decomposition (a vector of length \code{p*(p+1)/2}).  For \code{ilog_chol}, a variance matrix of size \code{p x p}.
}
\description{
Compute the log-Cholesky decomposition and its inverse.
}
\details{
The log-Cholesky decomposition of a variance matrix \code{V} is the upper triangular matrix, of which the non-zero elements are concatenated in column-major order.  Namely, the calculation is given by
\preformatted{
logC <- chol(V)
diag(logC) <- log(diag(logC))
logC <- logC[upper.tri(logC,diag=TRUE)]
}
}
