% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/losmix-package.R
\docType{package}
\name{losmix-package}
\alias{losmix}
\alias{losmix-package}
\title{losmix: Location-Scale Mixed-Effects Models}
\description{
Tools for fitting Gaussian linear mixed-effects models where both the mean and variance terms can be random.
}
\author{
\strong{Maintainer}: Martin Lysy \email{mlysy@uwaterloo.ca}

}
\keyword{internal}
