% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnix_pred.R
\name{mnix_pred}
\alias{mnix_pred}
\title{Genenerate random draws from the mNIX predictive distribution.}
\usage{
mnix_pred(n, Xpred, lambda, Omega, nu, tau, y, X, id)
}
\arguments{
\item{n}{Number of random draws (integer).}

\item{Xpred}{Covariate matrix of size \code{n x p} to use for the predictive distribution.}

\item{lambda}{Mean parameter.  A vector of length \code{p}, or a matrix of size \code{n x p}.}

\item{Omega}{Precision matrix parameter.  A matrix of size \code{p x p}, or an array of size \code{p x p x n}.}

\item{nu}{Degrees-of-freedom parameter.  A scalar or a vector of length \code{n}.}

\item{tau}{Scale parameter.  A scalar or a vector of length \code{n}.}

\item{y}{Optional response vector of length \code{N} (see \strong{Details}).}

\item{X}{Optional covariate matrix of size \code{N x p} (see \strong{Details}).}

\item{id}{Optional subject identifier vector of length \code{N} with \code{nsub <= N} distinct elements (see \strong{Details}).}
}
\value{
A vector of \code{n} draws from the predictive distribution.
}
\description{
Genenerate random draws from the mNIX predictive distribution.
}
\details{
If \code{y}, \code{X} and \code{id} are unspecified, then \code{(beta, sigma)} are drawn from \code{mNIX(lambda, Omega, nu, tau)}.  Otherwise, they are drawn from the posterior mNIX distribution \code{p(beta, sigma | y, X, id, lambda, Omega, nu, tau)}.  In this case, the number of subjects \code{nsub} must be \code{n} or \code{1}.
}
