% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnix_suff.R
\name{mnix_suff}
\alias{mnix_suff}
\title{Sufficient statistics for the mNIX distribution.}
\usage{
mnix_suff(id, y, X)
}
\arguments{
\item{id}{Subject identifiers.  A vector of length \code{n} consisting of \code{nsub <= n} distinct elements.}

\item{y}{Subject responses.  A vector of length \code{n}.}

\item{X}{Subject covariates.  A matrix of size \code{n x p}.}
}
\description{
Sufficient statistics for the mNIX distribution.
}
