% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnix_zeta.R
\name{mnix_zeta}
\alias{mnix_zeta}
\title{Calculate the mNIX normalizing constant.}
\usage{
mnix_zeta(Omega, nu, tau)
}
\arguments{
\item{Omega}{Precision matrix parameter.  A matrix of size \code{p x p}, or an array of size \code{p x p x n}.}

\item{nu}{Degrees-of-freedom parameter.  A scalar or a vector of length \code{n}.}

\item{tau}{Scale parameter.  A scalar or a vector of length \code{n}.}
}
\value{
Vector of mNIX normalizing constants.
}
\description{
Calculate the mNIX normalizing constant.
}
\details{
The mNIX normalzing constant (on the log scale) is defined as
\preformatted{
lgamma(nu/2) - .5 * ( nu * log(tau*nu/2) + log(|Omega|) )
}
}
