% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn.R
\name{mvn}
\alias{mvn}
\alias{rmvn}
\alias{dmvn}
\title{Density and random sampling for the multivariate normal distribution.}
\usage{
rmvn(n, mu, Sigma)

dmvn(x, mu, Sigma, log = FALSE)
}
\arguments{
\item{n}{Integer number of random draws.}

\item{mu}{Mean vector of length \code{p}.}

\item{Sigma}{Variance matrix of size \code{p x p}.}

\item{x}{Mutivariate observations.  A vector or length \code{n} or a matrix of size \code{n x p}.}

\item{log}{Logical; whether to calculate the density on the log scale.}
}
\value{
For \code{rmvn}, an \code{n x p} matrix of random draws.  For \code{dmvn} a vector of \code{n} density evaluations.
}
\description{
Density and random sampling for the multivariate normal distribution.
}
