% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{solveV}
\alias{solveV}
\title{Solve method for variance matrices.}
\usage{
solveV(V, b, ldV = FALSE)
}
\arguments{
\item{V}{Symmetric positive-definite (i.e., variance) matrix.}

\item{b}{Optional vector or matrix for which to solve the system of equations \code{V x = b}.  If missing calculates the inverse of \code{V}.}

\item{ldV}{Optionally compute the log determinant as well.}
}
\value{
The solution of the system of equations, or a list with elements \code{x} and \code{ldV} returning the log-determinant as well.
}
\description{
Solve method for variance matrices.
}
\details{
This function is faster and more stable than \code{base::solve} when \code{V} is known to be positive-definite.
}
