% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unlist_bind}
\alias{unlist_bind}
\title{Convert list of lists into list of arrays.}
\usage{
unlist_bind(x, name, bind)
}
\arguments{
\item{x}{List of (named) lists.}

\item{name}{Character vector of inner list elements to bind.  If missing defaults to every element of the inner list.}

\item{bind}{List of binding functions (e.g., \code{c}, \code{cbind}, etc.) the same length as \code{name}.  If missing defaults to \code{rep(c, length(name))}.}
}
\value{
A named list with elements corresponding to the binded values of the inner elements of \code{x}.
}
\description{
Convert list of lists into list of arrays.
}
\examples{
# generate a list of lists
Phi <- replicate(5, {
  list(lambda = rnorm(3), Omega = diag(rnorm(3)),
       nu = rexp(1), tau = rexp(1))
}, simplify = FALSE)

unlist_bind(Phi, bind = c(cbind, rbind, c, c))
}
