% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_phi.R
\name{vec_phi}
\alias{vec_phi}
\alias{ivec_phi}
\title{Convert between hyperparameter list and vector representation.}
\usage{
vec_phi(Phi)

ivec_phi(psi)
}
\arguments{
\item{Phi}{Hyperparameters as a list with elements \code{lambda}, \code{Omega}, \code{nu}, and \code{tau}.}

\item{psi}{Hyperparameters as an unconstrained numeric vector, or a matrix where each row is one hyperparameter set (see \strong{Details}).}
}
\value{
For \code{vec_phi}, an unconstrained vector or matrix.  For \code{ivec_phi}, a list with elements \code{lambda}, \code{Omega}, \code{nu}, and \code{tau}.
}
\description{
Convert between hyperparameter list and vector representation.
}
\details{
The unconstrained hyperparameter representation is given by
\preformatted{
psi <- c(lambda, log_chol(Omega), log(nu), log(tau))
}
where \code{\link{log_chol}} is the log-Cholesky decomposition.
}
