% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnix_marg.R
\name{mNIX_marg}
\alias{mNIX_marg}
\title{Marginal posterior of the mNIX hyperparameters as a \code{TMB::MakeADFun} object.}
\usage{
mNIX_marg(id, y, X)
}
\arguments{
\item{id}{Subject identifiers.  A vector of length \code{n} consisting of \code{nsub <= n} distinct elements.}

\item{y}{Subject responses.  A vector of length \code{n}.}

\item{X}{Subject covariates.  A matrix of size \code{n x p}.}
}
\description{
Marginal posterior of the mNIX hyperparameters as a \code{TMB::MakeADFun} object.
}
