% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnix_post.R
\name{mnix_post}
\alias{mnix_post}
\title{Calculate the hyperparameters of the mNIX conjugate posterior distribution.}
\usage{
mnix_post(id, y, X, lambda, Omega, tau, nu)
}
\arguments{
\item{id}{Subject identifiers.  A vector of length \code{n} consisting of \code{nsub <= n} distinct elements.}

\item{y}{Subject responses.  A vector of length \code{n}.}

\item{X}{Subject covariates.  A matrix of size \code{n x p}.}

\item{lambda}{Prior mean parameter.  A vector of length \code{p}, or a matrix of size \code{n x p}.}

\item{Omega}{Prior precision matrix.  A matrix of size \code{p x p}, or an array of size \code{p x p x n}.}

\item{tau}{Prior scale parameter.  A scalar or a vector of length \code{n}.}

\item{nu}{Prior degrees-of-freedom.  A scalar or a vector of length \code{n}.}
}
\description{
Calculate the hyperparameters of the mNIX conjugate posterior distribution.
}
