% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnix_sim.R
\name{mnix_sim}
\alias{mnix_sim}
\title{Genenerate random draws from the mNIX distribution.}
\usage{
mnix_sim(n, lambda, Omega, tau, nu, y, X)
}
\arguments{
\item{n}{Number of random draws (integer).}

\item{lambda}{Mean parameter.  A vector of length \code{p}, or a matrix of size \code{n x p}.}

\item{Omega}{Precision matrix parameter.  A matrix of size \code{p x p}, or an array of size \code{p x p x n}.}

\item{tau}{Scale parameter.  A scalar or a vector of length \code{n}.}

\item{nu}{Degrees-of-freedom parameter.  A scalar or a vector of length \code{n}.}

\item{y}{Optional response vector of length \code{N} for one subject.}

\item{X}{Optional covariate matrix of size \code{N x p} for one subject.}
}
\value{
A list with elements \code{beta} and \code{sigma} of size \code{n x p} and length \code{n}, respectively.
}
\description{
Genenerate random draws from the mNIX distribution.
}
