% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnix_suff.R
\name{mnix_suff}
\alias{mnix_suff}
\title{Sufficient statistics for the mNIX distribution.}
\usage{
mnix_suff(id, y, X)
}
\arguments{
\item{id}{Subject identifiers.  A vector of length \code{n} consisting of \code{nsub <= n} distinct elements.  If missing default to \code{rep(1, n)}, i.e., only one subject.}

\item{y}{Response vector for all subjects.  A vector of length \code{n}.}

\item{X}{Covariate matrix from all subjects.  A matrix of size \code{n x p}.}
}
\value{
A list with elements:
\describe{
  \item{\code{yy}}{A vector of length \code{nsub} consisting of \code{t(y) y} for each subject.}
  \item{\code{Xy}}{A matrix of size \code{p x nsub} consisting of \code{t(X) y} for each subject.}
  \item{\code{XX}}{A matrix of size \code{p x nsub^2} consisting of \code{t(X) X} for each subject.}
}
}
\description{
Sufficient statistics for the mNIX distribution.
}
